==> notebooks/3_model_training.ipynb <==
import io
import boto3
from dotenv import dotenv_values
import pandas as pd
import pandera as pa
from pandera import Column, DataFrameSchema, Check
from sklearn.model_selection import train_test_split
from pycaret.regression import (
    setup,
    finalize_model,
    save_model,
    predict_model,
)

# halfmarathon_df = pd.read_csv('../data/processed/halfmarathon_cleaned.csv')

env = dotenv_values(".env")

s3 = boto3.client(
    "s3",
    aws_access_key_id=env["AWS_ACCESS_KEY"],
    aws_secret_access_key=env["AWS_SECRET_KEY"],
    endpoint_url=env["AWS_ENDPOINT_URL"],
)

BUCKET_NAME = env["AWS_BUCKET_NAME"]

s3_key = f'data/processed/halfmarathon_cleaned.csv'
response = s3.get_object(Bucket=BUCKET_NAME, Key=s3_key)
halfmarathon_df = pd.read_csv(io.BytesIO(response['Body'].read()), sep=';')

halfmarathon_df.sample(5)
# halfmarathon_df.info()


schema = DataFrameSchema(
    {
        'gender': Column(
            pa.Int,
            checks=Check.in_range(0, 1),
            coerce=True,
            nullable=False
        ),
        'pace_5k_sec': Column(
            pa.Float,
            checks=Check.in_range(0, 765),
            coerce=True,
            nullable=True
        ),
        'time_5k_sec': Column(
            pa.Float,
            checks=Check.in_range(0, 3825),
            coerce=True,
            nullable=True
        ),
        'finish_time_sec': Column(
            pa.Int,
            checks=Check.in_range(3550, 10550),
            coerce=True,
            nullable=False
        ),
        'age': Column(
            pa.Int,
            checks=Check.in_range(6, 120),
            coerce=True,
            nullable=False
        )
    }
)

schema.validate(halfmarathon_df)

X = halfmarathon_df.drop('finish_time_sec', axis=1)
y = halfmarathon_df['finish_time_sec']

X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                    test_size=0.1,
                                                    random_state=99)

train_df = pd.concat([X_train, y_train], axis=1)
test_df = pd.concat([X_test, y_test], axis=1)

exp = setup(data=train_df, target='finish_time_sec' ,session_id=99, categorical_features=['gender'])

# normalized_exp = setup(data=halfmarathon_df, target='finish_time_sec' ,session_id=123, normalize=True, normalize_method='zscore',categorical_features=['gender'])

best_models = exp.compare_models(sort='MSE', n_select=5)
# best_models_normalized = exp.compare_models(sort='MAE', n_select=5)

tuned_best_models = [exp.tune_model(m, optimize='RMSE') for m in best_models]

best_model = exp.compare_models(best_models + tuned_best_models, sort="RMSE")

# exp.plot_model(best_model, plot='error')
exp.plot_model(best_model, plot='feature')
# exp.plot_model(best_model, plot='manifold')

best_final_model = exp.finalize_model(best_model)

predict_model(best_final_model, data=test_df)

final_train = pd.concat([train_df, test_df])

exp_all= setup(data=final_train, target='finish_time_sec', session_id=99)

final_model = finalize_model(best_model)

save_model(final_model, '../models/halfmarathon_predictor')
